+++ src/App.jsx
@@ -1,8 +1,10 @@
-import { useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
+import { DataTable } from "primereact/datatable";
 import { Column } from "primereact/column";
 import { Avatar } from "primereact/avatar";
-import { Tooltip } from "primereact/tooltip";
+import { InputText } from "primereact/inputtext";
 import { Header } from "./components/Header";
+import { FilterMatchMode } from "primereact/api";
 
 export default function App() {
   const [isLoading, setIsLoading] = useState(true);
@@ -33,6 +35,10 @@ export default function App() {
     return number.toLocaleString();
   };
 
+  const [filters, setFilters] = useState({
+    global: { value: null, matchMode: FilterMatchMode.CONTAINS },
+  });
+
   const getMedalEmoji = (rank) => {
     let emoji;
     switch (rank) {
@@ -51,66 +57,43 @@ export default function App() {
     return <span style={{ fontSize: "2em" }}>{emoji}</span>;
   };
 
-  const formatLastUpdated = () => {
-    if (!lastUpdated) {
-      return "";
-    }
-
-    const today = new Date();
-    const yesterday = new Date();
-    yesterday.setDate(yesterday.getDate() - 1);
-    const twoDaysAgo = new Date();
-    twoDaysAgo.setDate(twoDaysAgo.getDate() - 2);
-
-    const isToday = lastUpdated.toDateString() === today.toDateString();
-    const isYesterday = lastUpdated.toDateString() === yesterday.toDateString();
-    const isTwoDaysAgo =
-      lastUpdated.toDateString() === twoDaysAgo.toDateString();
-
-    const timeOptions = { hour: "2-digit", minute: "2-digit", hour12: false };
-    const formattedTime = lastUpdated.toLocaleTimeString(
-      undefined,
-      timeOptions
+  const renderHeader = () => {
+    return (
+      <div className="flex justify-content-end">
+        <span className="p-input-icon-left">
+          <i className="pi pi-search" />
+          <InputText
+            placeholder="Search Username"
+            onInput={(e) =>
+              setFilters({
+                global: {
+                  value: e.target.value,
+                  matchMode: FilterMatchMode.CONTAINS,
+                },
+              })
+            }
+          />
+        </span>
+      </div>
     );
-
-    if (isToday) {
-      return `Today at ${formattedTime} WIB`;
-    } else if (isYesterday) {
-      return `Yesterday at ${formattedTime} WIB`;
-    } else if (isTwoDaysAgo) {
-      return `2 Days Ago at ${formattedTime} WIB`;
-    } else {
-      return lastUpdated.toLocaleDateString();
-    }
   };
 
   return (
     <div className="card">
-      <Tooltip target=".custom-target-icon" />
-      <h1>IndoGitHubers</h1>
-      <h2>
-        Check Your GitHub Rank{" "}
-        <i
-          className="custom-target-icon pi pi-info-circle"
-          style={{ fontSize: "1.25rem" }}
-          data-pr-tooltip={`your GitHub account needs to have at least ${data.MinimumFollowerCount} followers to be on the list.`}
-          data-pr-position="right"
-        ></i>
-      </h2>
-      <h3>
-        Last Updated: {formatLastUpdated()}{" "}
-        <i
-          className="custom-target-icon pi pi-info-circle"
-          style={{ fontSize: "1rem" }}
-          data-pr-tooltip={`this data is not updated in real time. the data will be updated every 2 days.`}
-          data-pr-position="right"
-        ></i>
-      </h3>
+      <Header data={data} lastUpdated={lastUpdated} />
       {isLoading ? (
         <div>Loading...</div>
       ) : (
-        <>
-          <Header data={data.users} />
+        <DataTable
+          value={data.users}
+          tableStyle={{ minWidth: "50rem" }}
+          paginator
+          rows={10}
+          rowsPerPageOptions={[10, 25, 50]}
+          filters={filters}
+          header={renderHeader}
+          removableSort
+        >
           <Column
             field="avatarUrl"
             header="Name"
@@ -174,8 +157,7 @@ export default function App() {
             style={{ width: "30%" }}
             body={(rowData) => rowData.company || "-"}
           />
-        </>
-        // </DataTable>
+        </DataTable>
       )}
     </div>
   );
